# docs/static/swagger.json
swagger: '2.0'
info:
  title: CortexCommerce
  description: |2

    This API allows users to:
    - Browse products and categories
    - Manage shopping carts and orders
    - Apply discounts and process payments
    - Track order history and feedback
    All endpoints are secured where necessary and return JSON responses.
    For authentication, use JWT tokens
  license:
    name: BSD License
  version: v1
host: localhost:9000
schemes:
- http
basePath: /
consumes:
- application/json
produces:
- application/json
securityDefinitions:
  Bearer:
    type: apiKey
    name: Authorization
    in: header
    description: 'JWT Authorization header using the Bearer scheme. Example: ''Bearer
      {token}'''
security:
- Bearer: []
paths:
...
  /api/v1/payments/callback/:
    get:
      operationId: api_v1_payments_callback_list
      summary: Verify payment callback
      description: Verify payment with Zibal gateway using the given trackId
      parameters:
      - name: trackId
        in: query
        description: Unique transaction ID returned from Zibal
        required: true
        type: string
      responses:
        '200':
          description: Payment verification success
        '400':
          description: Payment failed or invalid trackId
        '502':
          description: Payment gateway connection error
      tags:
      - Payment
    parameters: []
...